Move Identity owner and delegate to new screen.
